# FRONTEND/src/contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fan Hub Plus API",
    "description": "Onion architecture API. Use the JWT accessToken returned by login as a Bearer token. All content types share /api/admin/content management routes.",
    "version": "v1"
  },
  "paths": {
...
    "/api/explore/articles": {
      "get": {
        "tags": ["Explore"],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Fandom",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Genre",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Tag",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "ReleaseYear",
            "in": "query",
            "schema": {
              "maximum": 2200,
              "minimum": 1900,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "maxItems": 8,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ContentType"
              }
            }
          },
          {
            "name": "ContentIds",
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "Upcoming",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "MinPopularity",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Featured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "pattern": "^(latest|popular|alphabetical|upcoming)$",
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDtoPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/explore/characters": {
      "get": {
        "tags": ["Explore"],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Fandom",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Genre",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Tag",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "ReleaseYear",
            "in": "query",
            "schema": {
              "maximum": 2200,
              "minimum": 1900,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "maxItems": 8,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ContentType"
              }
            }
          },
          {
            "name": "ContentIds",
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "Upcoming",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "MinPopularity",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Featured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "pattern": "^(latest|popular|alphabetical|upcoming)$",
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDtoPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/explore/videos": {
      "get": {
        "tags": ["Explore"],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Fandom",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Genre",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Tag",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "ReleaseYear",
            "in": "query",
            "schema": {
              "maximum": 2200,
              "minimum": 1900,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "maxItems": 8,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ContentType"
              }
            }
          },
          {
            "name": "ContentIds",
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "Upcoming",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "MinPopularity",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Featured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "pattern": "^(latest|popular|alphabetical|upcoming)$",
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDtoPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/explore/audio": {
      "get": {
        "tags": ["Explore"],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Fandom",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Genre",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Tag",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "ReleaseYear",
            "in": "query",
            "schema": {
              "maximum": 2200,
              "minimum": 1900,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "maxItems": 8,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ContentType"
              }
            }
          },
          {
            "name": "ContentIds",
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "Upcoming",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "MinPopularity",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Featured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "pattern": "^(latest|popular|alphabetical|upcoming)$",
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDtoPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/explore/galleries": {
      "get": {
        "tags": ["Explore"],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Fandom",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Genre",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Tag",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "ReleaseYear",
            "in": "query",
            "schema": {
              "maximum": 2200,
              "minimum": 1900,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "maxItems": 8,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ContentType"
              }
            }
          },
          {
            "name": "ContentIds",
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "Upcoming",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "MinPopularity",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Featured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "pattern": "^(latest|popular|alphabetical|upcoming)$",
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDtoPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/explore/merchandise": {
      "get": {
        "tags": ["Explore"],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Fandom",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Genre",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "Tag",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "ReleaseYear",
            "in": "query",
            "schema": {
              "maximum": 2200,
              "minimum": 1900,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "maxItems": 8,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ContentType"
              }
            }
          },
          {
            "name": "ContentIds",
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "Upcoming",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "MinPopularity",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Featured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "pattern": "^(latest|popular|alphabetical|upcoming)$",
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDtoPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/explore/upcoming": {
      "get": {
        "tags": ["Explore"],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDtoPage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "fandom": {
            "type": "string",
            "nullable": true
          },
          "genre": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "mediaUrl": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "featured": {
            "type": "boolean"
          },
          "views": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ticketUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ContentDtoPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentDto"
            },
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ContentType": {
        "enum": [
          "Article",
          "Character",
          "Video",
          "Audio",
          "Image",
          "Merchandise",
          "Release",
          "Event"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste the accessToken from POST /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}